Fixing cirricular dependencies and small updates
